Feature CRUD to API

+++ PetHolidays-API/database/migrations/2024_02_22_164832_create_accommodations_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->timestamps();
             $table->string('title');
-            $table->double('price', 8, 2);
+            $table->decimal('price', 8, 2);
             $table->string('city');
             $table->string('image');
             $table->string('description');
